+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 - Update kotlin to 1.6.20
+- Implement `MissingTestAnnotationTest`
 
 ## Version 0.0.3 *(2021-11-09)*
 - Fix a crash when using the plugin with the new detekt version 1.19.0-RC1 (https://github.com/detekt/detekt/issues/4228)